poppler: CVE-2024-6239
Issue: LIN1024-2375

A flaw was found in the Poppler's Pdfinfo utility. This issue
occurs when using -dests parameter with pdfinfo utility. By
using certain malformed input files, an attacker could cause
the utility to crash, leading to a denial of service.

CVE-2024-6239-0002 is the CVE fix and CVE-2024-6239-0001 is
dependent commit to fix the CVE.

Reference:
https://nvd.nist.gov/vuln/detail/CVE-2024-6239

Upstream patch:
https://gitlab.freedesktop.org/poppler/poppler/-/commit/0554731052d1a97745cb179ab0d45620589dd9c4
https://gitlab.freedesktop.org/poppler/poppler/-/commit/fc1c711cb5f769546c6b31cc688bf0ee7f0c1dbc

(LOCAL REV: NOT UPSTREAM) -- Will send to meta-openembedded

Signed-off-by: Yogita Urade <yogita.urade@windriver.com>