openvpn: remove obsolete patch to avoid do_patch error
Issue: LINUXEXEC-33650

Basically if the same patch is integrated into local repo and also
upstream repo, the patch contents will be kept if a merge do not
find conflicts. This is because even if the patch contents are the
same, they are treated as different commits.

(LOCAL REV: NOT UPSTREAM) - fix a merge issue

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>